Update how host tools get built
Bug: http://b/64121881

For Linux, build with glibc2.15 by mirroring soong's configuration for
the host (build/soong/cc/config/x86_linux_host.go).  Since we explicitly
specify a sysroot, Clang no longer includes the path to its
locally-installed libc++.so.  We have to explicitly specify it with the
linker flags.

Follow Soong configuration for Darwin as well, and use the sysroot
returned by 'xcrun --show-sdk-path'.  The minimum MacOS version is set
to 10.9 so libcxx can use ~typeinfo().

Turn off OpenMP for the host (which doesnot seem to be used anywhere),
as it is picking up some libc++ symbols inadvertently.

Test: build.py for Darwin and Linux, and test_compiler.py on Linux only.
Change-Id: I88230f49ee00460ea42d2bfb2ac97fedeedb6c35

Add Darwin support for hermetic builds.

Test: build.py on mac.

Change-Id: Ibeaf3b23b7e29c08ee6f9cea7c6e0fb7ecd84096