Use glibc from prebuilts for host Clang
Bug: http://b/64121881

Mirror soong configuration for the host
(build/soong/cc/config/x86_*_host.go) to build host Clang with
glibc2.15.

Since we explicitly specify a sysroot, Clang no longer includes the path
to its locally-installed libc++.so.  We have to explicitly specify it
with the linker flags.

Turn off OpenMP for the host (which does not seem to be used anywhere),
as it is picking up some libc++ symbols inadvertently.

Test: build.py for Darwin and Linux, and test_compiler.py on Linux only.
Change-Id: I88230f49ee00460ea42d2bfb2ac97fedeedb6c35