Fixes to build.py for LLVM update to r328903
Bug: http://b/77532044

- Use clang-4679922 for stage1 build

- Pass COMPILER_RT_FUZZER_LINK_LIBCXX=OFF to prevent libfuzzer from
linking in a private libcxx.  This is in conjunction with a local copy
of https://reviews.llvm.org/D44532.

- Turn off libfuzzer build during the native stage of Windows build.
This causes CMake to throw a multiply-defined-module error across the
cross-compilation and the native compilation (to build tablegen).

- It is no longer necessary to rename libclang.so.X.Y to libclang.so.X.
libclang.so.X.Y no longer exists - there's just libclang.so.X which has
the same name in the SONAME entry.

Test: build.py in llvm-toolchain branch.
Change-Id: I902f0257c0302e4cf20a8ff4f5957bbbac97da7b