Pass static-libgcc during windows build
Bug: http://b/112156620

With the recent MinGW update, GCC is built with --enable-shared so we
can get libgcc_eh.a (which is needed for building with Clang).  But in
this config, the way the static libraries are built, the linker prefers
to use symbols from the shared library instead of the static library.
In our setup, this'd imply carrying the shared libgcc*.DLL as extra
baggage.  We're instead using -static-libgcc to prefer the static
libraries (i.e. the old behavior) even if shared libgcc is available.

Test: clang.exe (32- and 64-bit) works under wine
Change-Id: Icff191ff452f6bb2728ca3217ae24bc8d09a72a8