Update patch set 1
Patch Set 1: Verified+1 Presubmit-Ready+1

This is just a reminder that we need to suppress this warning for now (because of skia). The update to aosp/master will also remove this at the build-system level, because we don't want to fix this issue actually (no bad C++11 compiler users in our platform).

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1