Update patch set 1
Patch Set 1: Code-Review-1

> Patch Set 1:
> 
> This may have to wait until we switch the default for the NDK over, or anyone not setting a linker will just get the current Android (non-lld) default. :(

Yeah, this changes the NDK defaults. That's still a few releases out.

If there's much value in doing this for the platform/kernel, we can work around this when installing Clang to the NDK by undoing it manually, but if this is just a clean up I'd prefer if we wait since I'd rather have as few special cases in our build as possible.

Patch-set: 1
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1