Add update_kernel_toolchain.py
This script automates most of my workflow for updating the version of
Clang used for various Linux kernel source trees used within Android.

It assumes that you:
1. Have a full kernel tree that's been checked out with `repo`.
2. Use revisions in `<kernel_tree>/.repo/manifests/default.xml` to "pin"
   the compiler revision.
3. Have a fresh checkout of AOSP.

Usage:
$ python update_kernel_toolchain.py /android0/kernel-wahoo \
    /android0/aosp/prebuilts/clang/host/linux-x86/clang-r365631b/bin/clang \
    141249463

This will use `/android/kernel-wahoo` as the kernel tree to
modify/update,
`/android0/aosp/prebuilts/clang/host/linux-x86/clang-r365631b/bin/clang`
as the binary of clang to upgrade to, and `141249463` as the bug number
to throw into the commit message.  The above will produce two commits,
and push them to Android gerrit (based on the remote, so potentially not
AOSP gerrit).  It will pair them together in a topic.  The two commits
can be the basis of a larger patch set, and should be manually boot
tested as well as presubmit tested.

Using a shell script that lists the different kernel trees, it's then
trivial to update many kernel tree's toolchain at once in an automated
fashion.

Bug: 141249463
Change-Id: I938792bd3991f860a46c8b99aea65060aea67d90
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>