Update patch set 3
Patch Set 3:

(3 comments)

Thanks for working on this shared wrapper. I'm happy to +2 it if anyone feels strongly about keeping line 1436 the same as you have it here. In addition to the git history side of it, I also don't think we use temp files very much in the build here, so it's kind of a departure from the keep everything and copy/package what we want as needed.

Patch-set: 3
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>