Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+2
> 
> I am a LLVM-Plugin developer. I use lib64/cmake/*.cmake in Android NDK. It works as my expect although there are no .a files, in fact I don't need .a files. 

We removed it because of changes to the NDK CMake support that cause our tests to regress when the CMake files are present.

How are the CMake files helpful in building the plugin?  Can you expand a bit?  Is it something you can configure manually?  The Clang/LLVM headers and the libraries are going to be in a standard location so you can depend on it to stay unchanged.

Patch-set: 2