Update patch set 2
Patch Set 2:

> Patch Set 2:
> > The Clang/LLVM headers and the libraries are going to be in a standard location so you can depend on it to stay unchanged.
> 
> Yes, building llvm plugin needs headers and libraries. But I think ${LLVM_DEFINITIONS} is also important(not sure), which might effect the build progress of llvm plugin.

We don't want the cmake files in the NDK, but it does seem reasonable to add it back to the prebuilts, since you're already downloading the headers from there.

danalbert@ Do you think the CMake subdirs can be pruned in the NDK import step instead of during the toolchain build?

Patch-set: 2