update_kernel_toolchain: work around 'repo info'
The 'repo info' doesn't reliably return the branch the user specified
with "repo init -b" if they reuse an existing checkout, causing the
script to push merges to the wrong remote.

Trust that "repo sync" did the right thing and use the merge branch
instead, which is the default for "repo upload".

Example of this problem:
$ repo info
Manifest branch: master
Manifest merge branch: refs/heads/common-android-mainline

Change-Id: I79381e4e2156b632284db124cb04772ce0b97d46