Add downstream changes as patch files
Bug: http://b/147699825

These are the complete set of patches for clang-r370808 prebuilts.

We are adopting the same format/layout as the ChromeOS toolchain (
https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/refs/heads/master/sys-devel/llvm/files/).

1. patches/PATCHES.json - json file listing the patches and the revision
numbers across which each patch is applicable.  The JSON entry for each
patch has start_revision/end_revision fields to hold this range.
2. patches/*.patch - downstream patches that never made it upstream.
3. patches/cherry/*.patch - changes that have a corresponding commit in
upstream.

Test: build.py with and without --build-llvm-next on the entire patch
stack.  When testing with --build-llvm-next, toolchain/llvm-project was
reset to 1bea97c971d60f261f1bdfaa7b6d9cb30a6962fd which corresponds to
r370808.

Change-Id: If653290ca2eb95e30301e1eeb071143b3b1eaca2