In a debug build, use stage1 compiler to build runtimes
Bug: http://b/149056576

In go/ab/aosp-llvm-toolchain, the linx_debug builds are taking 4 hours.
This is because runtimes are built with the debug compiler.  This is
slow on developer workstations as well.

Instead, stage1 to build the runtimes in this scenario.  To do this, we
should build the ARM and AArch64 backends during stage1.

With this change, I can get a debug build on my workstation in 1h5m.

Test: build.py --debug, presubmit
Change-Id: If42becf3f4af6d8e009989bece2957099e545f5b