Update patch set 2
Patch Set 2:

> Patch Set 2: Autosubmit+1 Verified+1
> 
> > Patch Set 2:
> > 
> > You can mark this Autosubmit, so it gets merged once it passes presubmit.  We can let it soak for a week or so to verify if the crashes remain.
> 
> Thanks! I was not sure if we ever want to submit it to verify. Will do

I was also a bit unclear.  Merging this CL will deploy the wrapper in the go/ab/aosp-master-plus-llvm build.  While this build uses the build servers, it does not affect any regular build or developer workflow.  We need to update the Clang prebuilts (b/149849077) for the new wrapper to be fully deployed.

The last time we submitted this change (with the glibc dependence), we reverted after seeing 4-5 breaks on this branch in a week. Let's hope it's no longer the case 😊

Patch-set: 2
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>