Update patch set 1
Patch Set 1: Code-Review+2

Looks good to me.

It looks like the platform build system pulls in [win]pthread in cc/linker.go:

	if ctx.Windows() {
		deps.LateStaticLibs = append(deps.LateStaticLibs, "libwinpthread")
	}

The MinGW libwinpthread.a and libpthread.a files are byte-for-byte identical.

$ cmp -bl prebuilts/gcc/linux-x86/host/x86_64-w64-mingw32-4.8/x86_64-w64-mingw32/lib/{libpthread.a,libwinpthread.a}
$ cmp -bl prebuilts/gcc/linux-x86/host/x86_64-w64-mingw32-4.8/x86_64-w64-mingw32/lib32/{libpthread.a,libwinpthread.a}

For Windows, the platform passes -nostdlib++ to the driver, but not -nostdlib.

Patch-set: 1
Reviewer: Gerrit User 1229339 <1229339@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2