Update patch set 2
Patch Set 2: Verified+1 Presubmit-Ready+1

Thanks, looks good to me. One thing I'd like to check with the build team is the risk of if this is enabled, NDK apps might come to use and then depend on it, so we might not be able to remove it in the future if needed.  This may be fine, but let me check with some other teams first before merging.

Patch-set: 2
Reviewer: Gerrit User 1114194 <1114194@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1