Retain host static libc++ libraries during packaging
In order to make it easier to use the Android toolchain as a drop-in
replacement on the host, we should keep the following static libraries:
- libc++.a
- libc++abi.a

We skip the libc++experimental.a library, since it actually has no
official packaging rule upstream either. If a user needs it, we can
provide it in the future.

Bug: http://b/152249445
Test: ./build.py and check for libc++*.a in lib64/.
Change-Id: I9091faf00cde251c74fe70ff119985b79758eb7c