Update llvm-next to r383902
* Set __STDC_FORMAT_MACROS for 32-bit host builds.

* 3ec3f62f0a0b cherry-picks cleanly now (and the backport doesn't).
Switch to the upstream version of this patch.  The exact version from
which the backport is unnecessary is somewhere betweek r380035 and
r383902 but we can find it later if necessary.

* Revert abb00753 and a2ca1c2d.  These two were reverted by upstream
change 916be8fd.  916be8fd also reverted two other changes 2181bf40 and
1079c68a, which are not in our current merge.
  start_revision: 380415 (that of a2ca1c2d).
  end_revision: 386220 (that of 916be8fd).
We may have to adjust the end revision if the next merge point is before
r386220.

Change-Id: I1c25521b31aff0c998099b85bddf5b476e3b5942