Adds a BuilderRegistry class to manage all build steps
(Any better naming suggestion?)

This class manages all built builders, so that others can find them when
necessary.

In the previous way, dependencies are on object but not the instance of
a builder. This makes it impossible to get any dynamic information. It
becomes a problem when I try to build runtime libs. E.g. the install_dir
may depends on arch, and not decidable statically.

Also it decides whether to skip a builder. I added --build and --skip options:

build.py                # build all
build.py --build a b c  # build a b and c but skip everything else
build.py --skip a b c   # skip a b and c but build everything else

Change-Id: I5ec7f458737ef2376836c1efb964a67b3789d159