Update patch set 3
Patch Set 3:

I'm not sure that this change was a good idea. If the goal of the bionic test is to test the bionic roundl() symbol (which is useful on its own because, assuming that the LLVM bug gets fixed eventually, NDK code compiled with future compilers will expect the symbol to behave correctly), then shouldn't the test be built with -ffreestanding? I think that should prevent the compiler assuming that it can replace roundl() calls with SSE instructions assuming round-to-nearest (and if it doesn't, that's a bug). And it would prevent further local changes like this as LLVM starts optimizing other fp math functions.

On the other hand, if the goal is to test "what happens when you write roundl() in your program", then the test as written is incorrect anyway because of a lack of fp_contract pragmas, and we can't add the pragmas because of the LLVM bug.

I would suggest adjusting the test to build with -ffreestanding, and then if/when the LLVM bug is fixed we can add a new test with the pragmas.

Patch-set: 3
CC: Gerrit User 1067098 <1067098@85c56323-6fa9-3386-8a01-6480fb634889>