Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > I'm not sure that this change was a good idea. If the goal of the bionic test is to test the bionic roundl() symbol (which is useful on its own because, assuming that the LLVM bug gets fixed eventually, NDK code compiled with future compilers will expect the symbol to behave correctly), then shouldn't the test be built with -ffreestanding? I think that should prevent the compiler assuming that it can replace roundl() calls with SSE instructions assuming round-to-nearest (and if it doesn't, that's a bug). And it would prevent further local changes like this as LLVM starts optimizing other fp math functions.
> > 
> > On the other hand, if the goal is to test "what happens when you write roundl() in your program", then the test as written is incorrect anyway because of a lack of fp_contract pragmas, and we can't add the pragmas because of the LLVM bug.
> > 
> > I would suggest adjusting the test to build with -ffreestanding, and then if/when the LLVM bug is fixed we can add a new test with the pragmas.
> 
> I don't think we want to build with -ffreestanding, because bionic is a libc. See https://github.com/android/ndk/issues/1117#issuecomment-551124817.

We'd be using -ffreestanding here to mean "don't make assumptions about the meaning of libc symbols", not "don't assume that libc is available". That's exactly what we want in general in the libc tests, because we don't want the compiler changing our "test this specific symbol" to "test some other symbol, or no symbol at all", as that may affect coverage.

> And changing the test and using this patch may cause app compatibility issue. So I prefer to wait until the pragma is available to see how to fix it.

It's only going to affect newly compiled apps, not existing ones. And those apps were broken anyway due to not using the pragma. If there's a real compatibility issue then it's not going to just affect Android but anything built using the new compiler, so upstream should be changed as well to be consistent.

Patch-set: 3