Add script to build kythe corpus
Bug: http://b/151096875

The script doesn't take any arguments and builds the kythe
cross-reference corpus.  First it builds toolchain and runtimes for the
host, directly from toolchain/llvm-project (i.e. not applying downstream
patches).  Then it invokes kythe's runextractor on the
compile_commands.json for that build.

There is one file where kythe's runextractor fails - mainly because the
driver cannot find the libcxx include path for a benchmark, which
actually gets generated at build time.

Package with gzip instead of bzip2.  The tradeoffs were:
    gzip: takes 2 mins, saves 120M
    bzip2: takes 13 mins, saves 230M

We can use pigz if it's part of Android build tools in the future.

Test: toolchain/llvm_android/kythe_xref.py
Change-Id: Idd4804e9009922c6f939716aa19ae54c960f6d88