Add script to build kythe corpus
Bug: http://b/151096875

The script builds toolchain and runtimes for the host, directly from
toolchain/llvm-project (i.e. not applying downstream patches).  Then it
invokes kythe's runextractor on the compile_commands.json.

There is one file where kythe's runextractor fails - mainly because the
driver cannot find the libcxx headers for a benchmark.  (The headers are
generated during build).

Package the kzip files using soong's merge_zips utility.  For this
reason this script needs to run in the master-plus-llvm branch.
Currently, the size of the generated corpus is 255M and the script runs
in 14minutes on my workstation:

real    14m3.375s
user    571m56.028s
sys     22m35.854s

Test: toolchain/llvm_android/kythe_xref.py
Change-Id: Idd4804e9009922c6f939716aa19ae54c960f6d88