Add test-prebuilts.py
Currently this script takes a build number and prepares a topic of CLs
in gerrit that can test the prebuilts on the Android platform.  Future
changes will add the ability to initiate Forrest runs on this topic.

gerrit.py has functionality to interact with gerrit and repo.

Test: test-prebuilts.py 6716542 created the topic
https: //r.android.com/q/topic:%22clang-testing-prebuilt-6716542%22
Change-Id: I3a439003d4f014f627505806a59b1c4c0d7ea28a