[update-prebuilts.py] Find branch using 'ab' command
Get the branch from the build server instead of guessing from the
current tree.

Also add a note to commit message of testing prebuilts.  This is an
added layer of warning to pervent merging them accidentally :)  (NB: If
the prebuilts are uploaded using the '--repo-upload option to this
script, they already have a '-2' from the uploader.)

Test: update-prebuilts.py with builds from release and testing branches.
Change-Id: If68557dc9ce315a160da8fec50a0b7c26da27142