[test scripts] Add functionality to invoke Forrest builds
test_prebuilts.py:
- Add a '--tag' parameter that can be used to group a set of
build/tests.
- Add functionality to invoke builds in Forrest.  (Functionality to
invoke tests will be added in the future.)
- The invocation IDs from Forrest are written to a forrest_pending.csv.
We're using a pending list because the presubmit build IDs are not
available at invocation time (and there may be more than one build in
case of infra error.)

update_results.py: New script to check  invocations in
forrest_pending.csv.  For invocations that have finished, find the build
ID and result and move the record to forrest.csv.

test_configs.json: JSON file with a list of platform configs we want to
test.
- Each config has a branch, target, a space-separated 'groups' property
and a list of tests.  List of tests is currently empty and will be added
in followups.
-  The groups property is used to annotate configs for easy
identification and filtering.  The filtering functionality will also be
added in followups.

Change-Id: I9c7a5b9cb01ea78d4cdaf54dae418076f832c3a6