[test/scripts] Submit tests in Forrest
Bug: http://b/161849352

Add functionality to submit tests on forrest.  Forrest has backend has
the ability to submit multiple tests but the forrest CLI tool doesn't.
So curate the Forrest request as a text protobuf and invoke it using
`forrest run <path-to-pb>`.

Add a cluster_info.yaml to store info needed for running tests on
Forrest.

Test: test_prebuilts.py --build 6842337 --tag 201012

Change-Id: I012db13bacbda0f11b31f8740fcadd0ae81e97f2