Set git remote to the AOSP project
LLVM build system infers the remote URL.  On the build servers, no
remote is set and the build uses a non-uniform path in /tmp.  Also, when
local mirrors are used, the remote URL points to the mirror.

While this path doesn't show up in the output of 'clang -v` it still
gets embedded in the binaries.  While setting up the sources in
$OUT/llvm-project, normalize the remote URL to the AOSP one (after
unlinking .git/config if it is a symlink).

Test: build.py and check stage1 output.
Change-Id: Iaa112b6c892a4985c74826a0949f36d3e23bb988