Output libunwind and link compiler-rt+libunwind into binaries
Update the toolchain build to use a new-style sysroot (from NDK r19 and
up). The old toplevel sysroot and platforms directories will be removed
in NDK r22. The new sysroot has the NDK libc++ STL in it already --
remove that for platform builders, but keep it for the NDK builders.

The libclang_rt.builtins-${arch}-android.a library is built for the NDK
and installed to both:
 - The Clang resource dir: stage2-install/lib64/clang/${ver}/lib/linux
 - the stage2-install/runtimes_ndk_cxx directory

Typically, the resource directory has the platform version of
compiler-rt libraries, but we only have a single stage2 Clang resource
directory for both NDK and platform builds, so we reuse the NDK library
on the platform. The NDK version should work for the platform, but the
platform version won't work for the NDK (e.g. older API levels have
`__sF` but not `stderr`).

Remove arm32 libunwind.a from the sysroot, and use a newly-built
libunwind.a for all architectures. libunwind.a is installed to the
Clang resource directory:
 - stage2-install/lib64/clang/${ver}/lib/linux/${arch}

Build libunwind.a using the "hermetic" option which marks the symbols
hidden. Also build a libunwind-exported.a w/o the hermetic flag, which
can be exported from libc.so in the platform.

lldb-server is a static binary built using libc.a from NDK r20, which
isn't quite good enough for the switch from libgcc.a to crt+libunwind.
For arm32 and x86-32, compile an object file from the current bionic
and insert it into libc.a:
 - On arm32, build bionic/exidx_static.c to provide
   dl_unwind_find_exidx that libunwind.a needs.
 - On x86-32, build bionic/__x86.get_pc_thunk.S to provide the
   __x86.get_pc_thunk.cx function that the libc.a string functions
   need.

The two libunwind cherrypicks disable the FrameHeaderCache. We would
like to leave that enabled, for the platform build only (not the NDK),
but the NDK r20 prebuilt link.h doesn't have the dlpi_adds/dlpi_subs
fields. They are added in r21. The libunwind _LIBUNWIND_IS_BAREMETAL
cherrypick avoids a merge conflict.

Bug: http://b/133247044
Test: build.py
Change-Id: I1f388dbfaf15c6c40c9a0a94b2f5ccef9226fa09